+++ ast/ppxlib_ast.ml
@@ -1,10 +1,15 @@
 open Import
 
+module type OCaml_version = Versions.OCaml_version
+
 module Ast            = Ast
 module Ast_helper     = Ast_helper
 module Ast_magic      = Selected_ast.Ast.Config
 module Asttypes       = Asttypes
-module Compiler_ast   = Versions.OCaml_current.Ast
+module Compiler_version = Versions.OCaml_current
+module Js             = Js
+module Find_version   = Versions.Find_version
+module Convert        = Versions.Convert
 module Extra_warnings = Warn
 module Lexer          = Lexer
 module Location_error = Location_error